+++ publish/CHANGELOG.md
@@ -1,3 +1,7 @@
+- v1.54
+  - Adds a new upgrade operation "deepnorth" to reset the Deep North and the ocean gap.
+  - Fixes the operation `zones_restore` not always working.
+
 - v1.53
   - Fixes the upgrade operation "ashlands" not resetting the ocean gap.